update swiftfin tvos signin
